Added default keymaps to build along with the customized ones

+++ build_firmwares.sh
@@ -8,6 +8,7 @@ echo "############################################"
 echo "#### Updating code from base repository ####"
 echo "############################################"
 git rebase upstream/master
+git push origin master -f
 
 # Build Signum 3.0 firmware
 echo "############################################"
@@ -19,6 +20,11 @@ if ! make signum/3_0/elitec:my_keymap;
 then
   exit 1
 fi
+make clean
+if ! make signum/3_0/elitec:default;
+then
+  exit 1
+fi
 
 # Build Crkbd rev1 firmware
 echo "############################################"
@@ -26,7 +32,12 @@ echo "####     Build Crkbd rev1 firmware      ####"
 echo "############################################"
 rm -rf .build
 make clean
-if ! make crkbd/rev1:my_keymap
+if ! make crkbd/rev1:my_keymap;
+then
+  exit 2
+fi
+make clean
+if ! make crkbd/rev1:default;
 then
   exit 2
 fi
@@ -37,18 +48,27 @@ echo "####      Build Iris rev4 firmware      ####"
 echo "############################################"
 rm -rf .build
 make clean
-if ! make keebio/iris/rev4:my_keymap
+if ! make keebio/iris/rev4:my_keymap;
 then
   exit 3
 fi
-
+make clean
+if ! make keebio/iris/rev4:default;
+then
+  exit 3
+fi
 # Build Quefrency rev1 firmware
 echo "############################################"
 echo "####   Build Quefrency rev1 firmware    ####"
 echo "############################################"
 rm -rf .build
 make clean
-if ! make keebio/quefrency/rev1:my_keymap
+if ! make keebio/quefrency/rev1:my_keymap;
+then
+  exit 4
+fi
+make clean
+if ! make keebio/quefrency/rev1:default;
 then
   exit 4
 fi
@@ -59,7 +79,12 @@ echo "####   Build Dactyl-Manuform firmware    ####"
 echo "#############################################"
 rm -rf .build
 make clean
-if ! make handwired/dactyl_manuform/5x6:my_keymap
+if ! make handwired/dactyl_manuform/5x6:my_keymap;
+then
+  exit 5
+fi
+make clean
+if ! make handwired/dactyl_manuform/5x6:default;
 then
   exit 5
 fi
